backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.1"
description = "srv webserver"
authors = ["Jay Gokhale <gokhale.jay@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9,<3.11"
python-dotenv = "^0.21.0"
rsa = "4.9"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
pydantic-collections = "^0.4.0"
email-validator = "^1.3.0"
statsd = "^4.0.1"
pymongo = "^4.3.3"
odmantic = "^0.9.1"
setuptools = "^67.2.0"
ordered-set = "^4.1.0"
jinja2 = "^3.1.2"
jira = "^3.5.0"
rich = "^13.3.4"
propelauth-fastapi = "^2.0.2"
fastapi-debug-toolbar = "^0.3.2"
gunicorn = "^20.1.0"
boto3 = "^1.26.135"
backoff = "^2.2.1"
resend = "^0.5.1"
posthog = "^3.0.1"
pdfkit = "^1.0.0"
tqdm = "^4.65.0"
python-multipart = "^0.0.6"
pygithub = "^1.58.2"
sentry-sdk = {extras = ["fastapi"], version = "^1.28.1"}
celery = "^5.3.1"
redis = "^4.6.0"
flower = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
python-magic = "^0.4.27"
httpx = "^0.25.0"

[tool.poetry.group.recalc.dependencies]
python = "^3.9,<3.11"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
